requirements.txt
altgraph==0.16.1
asgiref==3.3.1
blinker==1.4
boto==2.49.0
boto3==1.17.20
botocore==1.20.20
bottle==0.12.16
bottle-websocket==0.2.9
certifi==2017.7.27.1
cffi==1.12.3
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
django-cors-headers==3.2.1
django-environ==0.4.4
django-js-asset==1.1.0
django-mptt==0.9.0
django-mysql==2.4.1
django-ranged-response==0.2.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Eel==0.10.0
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.5
imageio==2.8.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.10.0
kiwisolver==1.0.1
libmagic==1.0
macholib==1.11
MarkupSafe==1.0
mysqlclient==1.3.10
passlib==1.7.1
pefile==2019.4.18
Pillow==5.2.0
pycparser==2.19
pycryptodome==3.9.7
pycryptodomex==3.9.7
PyInstaller==3.4
PyJWT==1.7.1
pymemcache==2.2.2
PyMySQL==0.7.11
pyparsing==2.4.0
pytesseract==0.3.3
python-dateutil==2.8.0
python-decouple==3.4
python-magic==0.4.15
python-magic-bin==0.4.14
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.5
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
PyYAML==5.1
requests==2.18.2
requests-oauthlib==0.8.0
s3transfer==0.3.4
schedule==0.6.0
scikit-image==0.16.2
scipy==1.4.1
shellescape==3.4.1
simplejson==3.17.0
six==1.10.0
sqlparse==0.4.1
unicode-slugify==0.1.3
Unidecode==1.0.23
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==15.1.0
Werkzeug==0.12.2
whichcraft==0.5.2
